requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0